requirements.txt
fastapi
pydantic
pandas
tensorflow
numpy
uvicorn
lightgbm
gdown
python-multipart
Pillow